tests/wpt/tests/html/canvas/element/fill-and-stroke-styles/2d.fillStyle.toStringFunctionCallback.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.fillStyle.toStringFunctionCallback</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.fillStyle.toStringFunctionCallback</h1>
<p class="desc">Passing a function in to ctx.fillStyle or ctx.strokeStyle with a toString callback works as specified</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Passing a function in to ctx.fillStyle or ctx.strokeStyle with a toString callback works as specified");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = { toString: function() { return "#008000"; } };
  _assertSame(ctx.fillStyle, "#008000", "ctx.fillStyle", "\"#008000\"");
  ctx.fillStyle = {};
  _assertSame(ctx.fillStyle, "#008000", "ctx.fillStyle", "\"#008000\"");
  ctx.fillStyle = 800000;
  _assertSame(ctx.fillStyle, "#008000", "ctx.fillStyle", "\"#008000\"");
  assert_throws_js(TypeError, function() { ctx.fillStyle = { toString: function() { throw new TypeError; } }; });
  ctx.strokeStyle = { toString: function() { return "#008000"; } };
  _assertSame(ctx.strokeStyle, "#008000", "ctx.strokeStyle", "\"#008000\"");
  ctx.strokeStyle = {};
  _assertSame(ctx.strokeStyle, "#008000", "ctx.strokeStyle", "\"#008000\"");
  ctx.strokeStyle = 800000;
  _assertSame(ctx.strokeStyle, "#008000", "ctx.strokeStyle", "\"#008000\"");
  assert_throws_js(TypeError, function() { ctx.strokeStyle = { toString: function() { throw new TypeError; } }; });

});
</script>
